paravirt/XEN/vsmp: Make paravirt thunks global
The paravirt thunks use a hack of using a static reference to a static
function to reference that function from the top level statement.

This assumes that gcc always generates static function names in a specific
format, which is not necessarily true.

Simply make these functions global and asmregparm. This way the
static __used variables are not needed and everything works.

Changed in paravirt and in all users (Xen and vsmp)

Signed-off-by: Andi Kleen <ak@linux.intel.com>