kbuild, kallsyms: Support single pass kallsyms
kallsyms currenly links the kernel upto three times
(in addition to another one for modpost checks)

Linking can be a quite slow operation, especially when
the kernel has a lot of debug information (lots of IO),
or Link Time Optimization is used.

Final linking is also a non parallelizable bottleneck, so it's
always good to do it less.

Use a different kallsyms method to avoid this:
- generate a initial kallsyms table from the top level
object files
- This table is usually a superset of the final table,
but without final addresses and some extra symbols
(e.g. discard and local symbols)
- Use this table to link the vmlinux
- Then generate a new kallsyms table with padding so
that all symbols stay at the same offsets. This works
because the new table is smaller or the same than the
original one.
We let the first kallsyms generate a padding file and
then use it on the next link.
- Then finally patch in the new table into the vmlinux

Right now we still do two links. One to generate
the kernel, and another one to generate the vmlinux.o
for modpost.

The kernel size is slightly increased from the padding.

On my slowish laptop this cuts down the final serialized phase
of a moderate size kernel build (just relinking vmlinux) by 1/3,
from ~30s to 20s. With LTO the wins are much higher.

Tested on x86, tests on other architectures would be appreciated.
This is still a somewhat experimental feature, so I made it
Kconfig. So far it is only enabled on x86, but can be enabled
elsewhere as testing progresses.

Signed-off-by: Andi Kleen <ak@linux.intel.com>