lto, workaround: Disable LTO for filter.c
Disable LTO for filter.c. This works around a gcc bug in the LTO
partitioner that partitions the jumptable from the BPF interpreter
into a different LTO unit. This in term causes assembler
errors because the jump table contains references to the
code labels in the original file.

Better workaround needed.

Signed-off-by: Andi Kleen <ak@linux.intel.com>