Avoid static variable in net/inet_frag.h
net/inet_frag.h has a static variable frag_percpu_counter_batch
which is used as a constant (like it was C++). Depending on the compiler
it can be duplicated multiple times, and also in C semantics it normally
cannot be eliminated.

Just replace it with a #define, since it only is a constant.

Signed-off-by: Andi Kleen <ak@linux.intel.com>