Allocate skb allocation function with __alloc_size
Add __alloc_size annotation to the skb allocation functions
with simple size arguments. Sums of arguments is not supported
by gcc, so those are not annotated.

Technically this will make the compiler assume the buffer size
applies to the __builtin_object_size of the skbuff header
(which is not correct). However since we shouldn't use
__builtin_object_size on skb headers (it is mainly used
for copy*user), we can still use __alloc_size and get
the additional warnings for bad value ranges.

Signed-off-by: Andi Kleen <ak@linux.intel.com>