Split KBUILD_VMLINUX_INIT
When building for PPC, the code from head_64.S should remain at the
start of the kernel, but LTO intermingles this with other symbols.

Since the code above refers to labels that are expected to be
relative to the start of the kernel (copy_to_here), the absolute
addresses generated by the assembler do not refer to the expected
code, and as a result, the platform fails to boot.

This patch splits KBUILD_VMLINUX_HEAD out of KBUILD_VMLINUX_INIT,
allowing LTO to be disabled for the head assembler files.

LTO is still usable for the rest of KBUILD_VMLINUX_INIT, as this
contains generic C code and no platform specific assembler.

Signed-off-by: Alastair D'Silva <alastair@d-silva.org>