delta: Fix buffer overrun in delta_ipc_open
delta_ipc_open is always called with a single constant string
as name, but it uses a longer memcpy to copy the string to
a different structure. The memcpy would read outside the bounds
of the string, potentially accessing unmapped memory.

Just use strcpy instead after clearing the area.

This fixes a build error with LTO, which can detect this.

Signed-off-by: Andi Kleen <ak@linux.intel.com>