crypto: Don't mark AES tables const and cacheline_aligned
cacheline_aligned is a special section. It cannot be const at the same
time because it's not read-only. It doesn't give any MMU protection.

If we wanted const cacheline aligned we would need to define
a new dedicated section. But I assume it's not really needed so the const
can be dropped.

This fixes LTO complaints about conflicting section types.

Signed-off-by: Andi Kleen <ak@linux.intel.com>