kbuild, genksyms: Support C output in genksyms
Support generating full C code for the CRC symbols instead of a
combination of inline assembler linker scripts.

That allows followon patches to avoid incremental linking.

This can be enabled with the new -c option.

This requires a new variant of the export keyword that also specifies
the section instead of just the symbol name, because that is needed
for the full symbol.

Signed-off-by: Andi Kleen <ak@linux.intel.com>