kbuild, modversions: Use C to generate genksyms crc symbols
The CRC symbols for genksym are currently generated by a combination of
inline assembler and a special linker script generated by genksyms,
and incremential linking for each exporting object.

In general we've been trying to minimize incremential linking. It can
be somewhat slow for large modules. Also it doesn't work with Link
Time Optimization in this case.

This patch cleans this all up by letting genksyms generate direct
C code for the CRC symbols, which then can be compiled directly
and linked in.

We generate a .ver.c file for each source file that exports symbols.

There are three cases how the objects can be used:

- ar'ed builtin. In this case we just cat the .ver.c files together,
build the object and include it in the archive.

- multi obj-m. These are linked together and also cat the .ver.c files
together and link them in.

- single file module. This case is a bit trickier. We generate a .ver.o
with the CRC symbols and rely on modpost to read that automatically
for the symbol checking. Then the .ver.o is linked in when the
final module is linked.

This all allows removing the inline assembler for generating the
CRC symbols, which is replaced by the generated CRC files.

Note the .ver.c files are currently implicit and not tracked through
dependencies.

The patch ends up being a net zero, as many deletions, as additions:

 6 files changed, 59 insertions(+), 59 deletions(-)

Signed-off-by: Andi Kleen <ak@linux.intel.com>