kbuild, modpost: Read .ver.o. files to resolve CRC symbols
Followon patches will convert the Makefiles to use C to generate
the MODVERSIONS CRC symbols. Normally these symbols are linked or
ar'ed in, but for the case of a single file module we don't
want to use incremental linking just to link in the CRC symbol.
In this case modpost runs on the original object, which is missing
the CRC symbols and would complain about it.

Add support for automatically checking for .ver.o files and reading
them for the CRC symbol for each module. This way we can avoid
an extra link step just for modpost checking.

The CRC symbols are then really linked in in the final module
linking step after modpost.

Signed-off-by: Andi Kleen <ak@linux.intel.com>