kbuild, kallsyms: Support compiling relative_base separately
The patching method used by single link kallsyms cannot handle
relocations. They always get lost during the objcopy to a binary.

There is a single relocation in a kallsyms file for the
kallsyms_relative_base, normally pointing to _text.

Add some ifdefs to the kallsyms output that allows compiling
kallsyms_relative_base in a separate object file. The base
does not change between kallsyms passes, so it can be always
linked in from the initial version.

Signed-off-by: Andi Kleen <ak@linux.intel.com>