sys_ni: Switch SYSCALL_ALIAS/cond_syscall to use C alias declarations
Now that we have prototypes for everything used in sys_ni.c
we can switch the alias macros over to use C declarations.

For this we need to include compat.h and syscalls.h in sys_ni.c
(there was a comment that this wasn't possible, but it's now)

This avoids needing to use assembler for this and is needed
for gcc LTO.

I kept the architecture specific alias macros for now, even
though they are likely not needed any more.

One problem is that the compiler warns if the alias prototypes
do not match. This is unfortunately common for compat prototypes.
Just disable that warning for sys_ni.c

Signed-off-by: Andi Kleen <ak@linux.intel.com>