btf: Fix up btf_ids.h for gcc lto
gcc lto doesn't guarantee that top level asm() statements are in a
particular order, and also they might end up in different assembler
files.

This causes problems with the way the btf ids are generated, which
assumes a hole less placement of asm statements.

Make all the symbols global so that we avoid linker errors
if the asm statement is somewhere else.

For BTF list use a single asm() statement through the macros,
to keep them all together.

Probably it would be better to rewrite some of this in C, then
we could at least use __noreorder.

Cc: jolsa@kernel.org
Signed-off-by: Andi Kleen <andi@firstfloor.org>